refactor(evmd): move mempool wiring to its own file, fix changelog PR link
Addresses review: keep app.go free of mempool-specific logic and point
the changelog entry at this PR (#1051) instead of the closed #1034.
